Fixed 4k images with many people not rendering skeletons

+++ src/openpose/hand/handGpuRenderer.cpp
@@ -84,7 +84,7 @@ namespace op
                 // I prefer std::round(T&) over positiveIntRound(T) for std::atomic
                 const auto elementRendered = spElementToRender->load();
                 const auto numberPeople = handKeypoints[0].getSize(0);
-                const Point<int> frameSize{outputData.getSize(1), outputData.getSize(0)};
+                const Point<unsigned int> frameSize{(unsigned int)outputData.getSize(1), (unsigned int)outputData.getSize(0)};
                 // GPU rendering
                 if (numberPeople > 0 && elementRendered == 0)
                 {
